fix: fixed a redirect bug from sign in to game
